Modularize codebase into separate source files
Split the monolithic OutputParser.swift (1500+ lines) into focused modules:

- Models.swift: Data structures (BuildResult, BuildSummary, BuildError,
  BuildWarning, FailedTest, CodeCoverage, FileCoverage) and GitHub Actions
  formatting logic
- CoverageParser.swift: Coverage file parsing, auto-detection, and
  conversion (.profraw/.xcresult to JSON)
- OutputParser.swift: Build output parsing only (errors, warnings, tests,
  build time extraction)

This improves maintainability by giving each file a single responsibility
and makes the codebase easier to navigate and test independently.
