+++ Sources/OutputParser.swift
@@ -1,282 +1,6 @@
 import Foundation
 import RegexBuilder
 
-struct BuildResult: Codable {
-    let status: String
-    let summary: BuildSummary
-    let errors: [BuildError]
-    let warnings: [BuildWarning]
-    let failedTests: [FailedTest]
-    let coverage: CodeCoverage?
-    let printWarnings: Bool
-    let printCoverageDetails: Bool
-
-    enum CodingKeys: String, CodingKey {
-        case status, summary, errors, warnings, coverage
-        case failedTests = "failed_tests"
-    }
-
-    init(
-        status: String,
-        summary: BuildSummary,
-        errors: [BuildError],
-        warnings: [BuildWarning],
-        failedTests: [FailedTest],
-        coverage: CodeCoverage?,
-        printWarnings: Bool,
-        printCoverageDetails: Bool = false
-    ) {
-        self.status = status
-        self.summary = summary
-        self.errors = errors
-        self.warnings = warnings
-        self.failedTests = failedTests
-        self.coverage = coverage
-        self.printWarnings = printWarnings
-        self.printCoverageDetails = printCoverageDetails
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        status = try container.decode(String.self, forKey: .status)
-        summary = try container.decode(BuildSummary.self, forKey: .summary)
-        errors = try container.decodeIfPresent([BuildError].self, forKey: .errors) ?? []
-        warnings = try container.decodeIfPresent([BuildWarning].self, forKey: .warnings) ?? []
-        failedTests = try container.decodeIfPresent([FailedTest].self, forKey: .failedTests) ?? []
-        coverage = try container.decodeIfPresent(CodeCoverage.self, forKey: .coverage)
-        printWarnings = false
-        printCoverageDetails = false
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(status, forKey: .status)
-        try container.encode(summary, forKey: .summary)
-
-        if !errors.isEmpty {
-            try container.encode(errors, forKey: .errors)
-        }
-
-        if printWarnings && !warnings.isEmpty {
-            try container.encode(warnings, forKey: .warnings)
-        }
-
-        if !failedTests.isEmpty {
-            try container.encode(failedTests, forKey: .failedTests)
-        }
-
-        // Only output coverage section in details mode
-        // In summary-only mode, coverage_percent in summary is sufficient
-        if let coverage = coverage, printCoverageDetails {
-            try container.encode(coverage, forKey: .coverage)
-        }
-    }
-
-    // MARK: - GitHub Actions Output
-
-    /// Formats the build result as GitHub Actions workflow commands
-    func formatGitHubActions() -> String {
-        var output: [String] = []
-
-        // Format errors as ::error commands
-        for error in errors {
-            output.append(formatGitHubActionsError(error))
-        }
-
-        // Format warnings as ::warning commands
-        if printWarnings {
-            for warning in warnings {
-                output.append(formatGitHubActionsWarning(warning))
-            }
-        }
-
-        // Format failed tests as ::error commands
-        for test in failedTests {
-            output.append(formatGitHubActionsTest(test))
-        }
-
-        // Add summary notice
-        let summaryMessage = buildSummaryMessage()
-        output.append("::notice ::\(summaryMessage)")
-
-        return output.joined(separator: "\n")
-    }
-
-    private func formatGitHubActionsError(_ error: BuildError) -> String {
-        let fileComponents = formatFileComponents(file: error.file, line: error.line, column: error.column)
-        return "::\("error") \(fileComponents)::\(error.message)"
-    }
-
-    private func formatGitHubActionsWarning(_ warning: BuildWarning) -> String {
-        let fileComponents = formatFileComponents(file: warning.file, line: warning.line, column: warning.column)
-        return "::\("warning") \(fileComponents)::\(warning.message)"
-    }
-
-    private func formatGitHubActionsTest(_ test: FailedTest) -> String {
-        var fileComponents = formatFileComponents(file: test.file, line: test.line, column: test.column)
-        // Add test name as title for better visibility in GitHub Actions
-        if !fileComponents.isEmpty {
-            fileComponents += ","
-        }
-        fileComponents += "title=\(test.test)"
-        return "::\("error") \(fileComponents)::\(test.message)"
-    }
-
-    private func formatFileComponents(file: String?, line: Int?, column: Int?) -> String {
-        guard let file = file else {
-            return ""
-        }
-
-        guard let line = line else {
-            return "file=\(file)"
-        }
-
-        if let column = column {
-            return "file=\(file),line=\(line),col=\(column)"
-        }
-
-        return "file=\(file),line=\(line)"
-    }
-
-    private func buildSummaryMessage() -> String {
-        var parts: [String] = []
-
-        if status == "success" {
-            parts.append("Build succeeded")
-        } else {
-            parts.append("Build failed")
-        }
-
-        if summary.errors > 0 {
-            parts.append("\(summary.errors) error\(summary.errors == 1 ? "" : "s")")
-        }
-
-        if summary.warnings > 0 {
-            parts.append("\(summary.warnings) warning\(summary.warnings == 1 ? "" : "s")")
-        }
-
-        if summary.failedTests > 0 {
-            parts.append("\(summary.failedTests) failed test\(summary.failedTests == 1 ? "" : "s")")
-        }
-
-        if let passedTests = summary.passedTests, passedTests > 0 {
-            parts.append("\(passedTests) passed test\(passedTests == 1 ? "" : "s")")
-        }
-
-        if let buildTime = summary.buildTime {
-            parts.append("in \(buildTime)")
-        }
-
-        if let coveragePercent = summary.coveragePercent {
-            parts.append(String(format: "%.1f%% coverage", coveragePercent))
-        }
-
-        return parts.joined(separator: ", ")
-    }
-}
-
-struct BuildSummary: Codable {
-    let errors: Int
-    let warnings: Int
-    let failedTests: Int
-    let passedTests: Int?
-    let buildTime: String?
-    let coveragePercent: Double?
-
-    enum CodingKeys: String, CodingKey {
-        case errors
-        case warnings
-        case failedTests = "failed_tests"
-        case passedTests = "passed_tests"
-        case buildTime = "build_time"
-        case coveragePercent = "coverage_percent"
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(errors, forKey: .errors)
-        try container.encode(warnings, forKey: .warnings)
-        try container.encode(failedTests, forKey: .failedTests)
-
-        // Only encode optional fields if they have values
-        if let passedTests = passedTests {
-            try container.encode(passedTests, forKey: .passedTests)
-        }
-        if let buildTime = buildTime {
-            try container.encode(buildTime, forKey: .buildTime)
-        }
-        if let coveragePercent = coveragePercent {
-            try container.encode(coveragePercent, forKey: .coveragePercent)
-        }
-    }
-}
-
-struct BuildError: Codable {
-    let file: String?
-    let line: Int?
-    let message: String
-
-    // Internal only - used for GitHub Actions format, not encoded to JSON/TOON
-    var column: Int? = nil
-
-    enum CodingKeys: String, CodingKey {
-        case file, line, message
-    }
-}
-
-struct BuildWarning: Codable {
-    let file: String?
-    let line: Int?
-    let message: String
-
-    // Internal only - used for GitHub Actions format, not encoded to JSON/TOON
-    var column: Int? = nil
-
-    enum CodingKeys: String, CodingKey {
-        case file, line, message
-    }
-}
-
-struct FailedTest: Codable {
-    let test: String
-    let message: String
-    let file: String?
-    let line: Int?
-
-    // Internal only - used for GitHub Actions format, not encoded to JSON/TOON
-    var column: Int? = nil
-
-    enum CodingKeys: String, CodingKey {
-        case test, message, file, line
-    }
-}
-
-struct CodeCoverage: Codable {
-    let lineCoverage: Double
-    let files: [FileCoverage]
-
-    enum CodingKeys: String, CodingKey {
-        case lineCoverage = "line_coverage"
-        case files
-    }
-}
-
-struct FileCoverage: Codable {
-    let path: String
-    let name: String
-    let lineCoverage: Double
-    let coveredLines: Int
-    let executableLines: Int
-
-    enum CodingKeys: String, CodingKey {
-        case path
-        case name
-        case lineCoverage = "line_coverage"
-        case coveredLines = "covered_lines"
-        case executableLines = "executable_lines"
-    }
-}
-
 class OutputParser {
     private var errors: [BuildError] = []
     private var warnings: [BuildWarning] = []
@@ -1083,439 +807,5 @@ class OutputParser {
         }
 
         return nil
-    }
-
-    // MARK: - Code Coverage Parsing
-
-    // MARK: - Coverage Auto-Conversion Helpers
-
-    /// Runs a shell command and returns the output
-    private static func runShellCommand(_ command: String, args: [String]) -> String? {
-        let process = Process()
-        process.executableURL = URL(fileURLWithPath: "/usr/bin/env")
-        process.arguments = [command] + args
-
-        let pipe = Pipe()
-        process.standardOutput = pipe
-        process.standardError = pipe
-
-        do {
-            try process.run()
-
-            // Read data BEFORE waiting for exit to avoid pipe deadlock
-            let data = pipe.fileHandleForReading.readDataToEndOfFile()
-            process.waitUntilExit()
-
-            guard process.terminationStatus == 0 else {
-                return nil
-            }
-
-            return String(data: data, encoding: .utf8)
-        } catch {
-            return nil
-        }
-    }
-
-    /// Finds all .profraw files in a directory (recursively)
-    private static func findProfrawFiles(in directory: String) -> [String] {
-        let fileManager = FileManager.default
-        guard let enumerator = fileManager.enumerator(atPath: directory) else {
-            return []
-        }
-
-        var profrawFiles: [String] = []
-        for case let file as String in enumerator {
-            if file.hasSuffix(".profraw") {
-                let fullPath = (directory as NSString).appendingPathComponent(file)
-                profrawFiles.append(fullPath)
-            }
-        }
-        return profrawFiles
-    }
-
-    /// Finds the test binary (.xctest bundle) in .build directory
-    private static func findTestBinary() -> String? {
-        let fileManager = FileManager.default
-        let buildDir = ".build"
-
-        guard fileManager.fileExists(atPath: buildDir),
-            let enumerator = fileManager.enumerator(atPath: buildDir)
-        else {
-            return nil
-        }
-
-        for case let file as String in enumerator {
-            if file.hasSuffix(".xctest") {
-                let xctestPath = (buildDir as NSString).appendingPathComponent(file)
-                let macosPath = (xctestPath as NSString).appendingPathComponent("Contents/MacOS")
-
-                guard let macosContents = try? fileManager.contentsOfDirectory(atPath: macosPath) else {
-                    continue
-                }
-
-                for item in macosContents {
-                    let itemPath = (macosPath as NSString).appendingPathComponent(item)
-                    var isDirectory: ObjCBool = false
-
-                    if fileManager.fileExists(atPath: itemPath, isDirectory: &isDirectory),
-                        !isDirectory.boolValue,
-                        !item.hasSuffix(".dSYM")
-                    {
-                        return itemPath
-                    }
-                }
-            }
-        }
-        return nil
-    }
-
-    /// Converts .profraw files to JSON coverage data
-    private static func convertProfrawToJSON(profrawFiles: [String]) -> CodeCoverage? {
-        guard !profrawFiles.isEmpty else {
-            return nil
-        }
-
-        guard let testBinary = findTestBinary() else {
-            return nil
-        }
-
-        let tempDir = NSTemporaryDirectory()
-        let profdataPath = (tempDir as NSString).appendingPathComponent("xcsift-coverage.profdata")
-        let jsonPath = (tempDir as NSString).appendingPathComponent("xcsift-coverage.json")
-
-        let mergeArgs = ["llvm-profdata", "merge", "-sparse"] + profrawFiles + ["-o", profdataPath]
-        guard runShellCommand("xcrun", args: mergeArgs) != nil else {
-            return nil
-        }
-
-        let exportArgs = ["llvm-cov", "export", testBinary, "-instr-profile=\(profdataPath)", "-format=text"]
-        guard let jsonOutput = runShellCommand("xcrun", args: exportArgs) else {
-            try? FileManager.default.removeItem(atPath: profdataPath)
-            return nil
-        }
-
-        guard let jsonData = jsonOutput.data(using: .utf8) else {
-            try? FileManager.default.removeItem(atPath: profdataPath)
-            return nil
-        }
-
-        do {
-            try jsonData.write(to: URL(fileURLWithPath: jsonPath))
-            let coverage = parseCoverageJSON(at: jsonPath)
-            try? FileManager.default.removeItem(atPath: profdataPath)
-            try? FileManager.default.removeItem(atPath: jsonPath)
-            return coverage
-        } catch {
-            try? FileManager.default.removeItem(atPath: profdataPath)
-            return nil
-        }
-    }
-
-    /// Finds .xcresult bundles (created by xcodebuild)
-    private static func findXCResultBundles(in directory: String) -> [String] {
-        let fileManager = FileManager.default
-        guard let enumerator = fileManager.enumerator(atPath: directory) else {
-            return []
-        }
-
-        var xcresultPaths: [String] = []
-        for case let file as String in enumerator {
-            if file.hasSuffix(".xcresult") {
-                let fullPath = (directory as NSString).appendingPathComponent(file)
-                xcresultPaths.append(fullPath)
-            }
-        }
-        return xcresultPaths
-    }
-
-    /// Finds the most recent .xcresult bundle in Xcode DerivedData using shell find
-    private static func findLatestXCResultInDerivedData(projectHint: String? = nil) -> String? {
-        let fileManager = FileManager.default
-        let homeDir = fileManager.homeDirectoryForCurrentUser.path
-        let derivedDataPath = (homeDir as NSString).appendingPathComponent("Library/Developer/Xcode/DerivedData")
-
-        guard fileManager.fileExists(atPath: derivedDataPath) else {
-            return nil
-        }
-
-        // If we have a project hint, search only in matching project directories
-        let searchPaths: [String]
-        if let hint = projectHint {
-            // Find project directories matching the hint (e.g., "MyProject-abcdef123")
-            let projectDirs = (try? fileManager.contentsOfDirectory(atPath: derivedDataPath))?
-                .filter { $0.hasPrefix("\(hint)-") || $0.hasPrefix("\(hint)Tests-") }
-                .map { (derivedDataPath as NSString).appendingPathComponent($0) }
-                .filter { fileManager.fileExists(atPath: $0) }
-
-            guard let dirs = projectDirs, !dirs.isEmpty else {
-                // Project hint provided but no matching directory found
-                // Return nil to allow fallback to SPM paths
-                return nil
-            }
-            searchPaths = dirs
-        } else {
-            searchPaths = [derivedDataPath]
-        }
-
-        let findArgs = ["find"] + searchPaths + ["-name", "*.xcresult", "-type", "d", "-mtime", "-7"]
-        guard let output = runShellCommand("/usr/bin/env", args: findArgs) else {
-            return nil
-        }
-
-        let paths = output.components(separatedBy: .newlines).filter { !$0.isEmpty }
-        guard !paths.isEmpty else {
-            return nil
-        }
-
-        var newestBundle: String?
-        var newestDate: Date?
-
-        for path in paths {
-            if let attrs = try? fileManager.attributesOfItem(atPath: path),
-                let modDate = attrs[.modificationDate] as? Date
-            {
-                if newestDate == nil || modDate > newestDate! {
-                    newestDate = modDate
-                    newestBundle = path
-                }
-            }
-        }
-
-        return newestBundle
-    }
-
-    /// Converts .xcresult bundle to JSON coverage data
-    private static func convertXCResultToJSON(xcresultPath: String, targetFilter: String? = nil) -> CodeCoverage? {
-        let args = ["xccov", "view", "--report", "--json", xcresultPath]
-        guard let jsonOutput = runShellCommand("xcrun", args: args) else {
-            return nil
-        }
-
-        guard let jsonData = jsonOutput.data(using: .utf8),
-            let json = try? JSONSerialization.jsonObject(with: jsonData) as? [String: Any]
-        else {
-            return nil
-        }
-
-        return parseXcodebuildFormat(json: json, targetFilter: targetFilter)
-    }
-
-    static func parseCoverageFromPath(_ path: String, targetFilter: String? = nil) -> CodeCoverage? {
-        let fileManager = FileManager.default
-        let coveragePath: String
-
-        // If explicit path provided and exists, use it
-        if !path.isEmpty && fileManager.fileExists(atPath: path) {
-            coveragePath = path
-        } else {
-            // Auto-detect: try xcodebuild first (DerivedData), then SPM paths
-            if let latestXCResult = findLatestXCResultInDerivedData(projectHint: targetFilter) {
-                return convertXCResultToJSON(xcresultPath: latestXCResult, targetFilter: targetFilter)
-            }
-
-            let defaultPaths = [
-                ".build/debug/codecov",
-                ".build/arm64-apple-macosx/debug/codecov",
-                ".build/x86_64-apple-macosx/debug/codecov",
-                ".build/arm64-unknown-linux-gnu/debug/codecov",
-                ".build/x86_64-unknown-linux-gnu/debug/codecov",
-                "DerivedData",
-                ".",
-            ]
-
-            var foundPath: String?
-            for defaultPath in defaultPaths {
-                if fileManager.fileExists(atPath: defaultPath) {
-                    foundPath = defaultPath
-                    break
-                }
-            }
-
-            guard let found = foundPath else {
-                return nil
-            }
-            coveragePath = found
-        }
-
-        guard fileManager.fileExists(atPath: coveragePath) else {
-            return nil
-        }
-
-        var isDirectory: ObjCBool = false
-        _ = fileManager.fileExists(atPath: coveragePath, isDirectory: &isDirectory)
-
-        if isDirectory.boolValue {
-            guard let files = try? fileManager.contentsOfDirectory(atPath: coveragePath) else {
-                return nil
-            }
-
-            let jsonFiles = files.filter { $0.hasSuffix(".json") }
-            if let firstJsonFile = jsonFiles.first {
-                let jsonPath = (coveragePath as NSString).appendingPathComponent(firstJsonFile)
-                return parseCoverageJSON(at: jsonPath, targetFilter: targetFilter)
-            }
-
-            let profrawFiles = findProfrawFiles(in: coveragePath)
-            if !profrawFiles.isEmpty {
-                return convertProfrawToJSON(profrawFiles: profrawFiles)
-            }
-
-            let xcresultBundles = findXCResultBundles(in: coveragePath)
-            if let firstXCResult = xcresultBundles.first {
-                return convertXCResultToJSON(xcresultPath: firstXCResult, targetFilter: targetFilter)
-            }
-
-            if let latestXCResult = findLatestXCResultInDerivedData() {
-                return convertXCResultToJSON(xcresultPath: latestXCResult, targetFilter: targetFilter)
-            }
-
-            return nil
-        } else {
-            if coveragePath.hasSuffix(".xcresult") {
-                return convertXCResultToJSON(xcresultPath: coveragePath, targetFilter: targetFilter)
-            } else {
-                return parseCoverageJSON(at: coveragePath, targetFilter: targetFilter)
-            }
-        }
-    }
-
-    private static func parseCoverageJSON(at path: String, targetFilter: String? = nil) -> CodeCoverage? {
-        guard let data = try? Data(contentsOf: URL(fileURLWithPath: path)) else {
-            return nil
-        }
-
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            return nil
-        }
-
-        if let coverage = parseXcodebuildFormat(json: json, targetFilter: targetFilter) {
-            return coverage
-        }
-
-        if let coverage = parseSPMFormat(json: json) {
-            return coverage
-        }
-
-        return nil
-    }
-
-    private static func parseXcodebuildFormat(json: [String: Any], targetFilter: String? = nil) -> CodeCoverage? {
-        guard let targets = json["targets"] as? [[String: Any]] else {
-            return nil
-        }
-
-        var fileCoverages: [FileCoverage] = []
-        var totalCovered = 0
-        var totalExecutable = 0
-
-        for target in targets {
-            // Get target name if available
-            let targetName = target["name"] as? String
-
-            // Skip test bundles - we want coverage of tested code, not tests themselves
-            if let name = targetName, name.hasSuffix(".xctest") {
-                continue
-            }
-
-            // Apply target filter if specified
-            if let filter = targetFilter, let name = targetName {
-                if !name.contains(filter) && !filter.contains(name) {
-                    continue
-                }
-            }
-
-            guard let filesArray = target["files"] as? [[String: Any]] else {
-                continue
-            }
-
-            for fileData in filesArray {
-                guard let filename = fileData["path"] as? String else {
-                    continue
-                }
-
-                let lineCoverage: Double
-                let covered: Int
-                let executable: Int
-
-                if let coverage = fileData["lineCoverage"] as? Double {
-                    lineCoverage = coverage > 1.0 ? coverage : coverage * 100.0
-                    covered = fileData["coveredLines"] as? Int ?? 0
-                    executable = fileData["executableLines"] as? Int ?? 0
-                } else {
-                    continue
-                }
-
-                let name = (filename as NSString).lastPathComponent
-
-                fileCoverages.append(
-                    FileCoverage(
-                        path: filename,
-                        name: name,
-                        lineCoverage: lineCoverage,
-                        coveredLines: covered,
-                        executableLines: executable
-                    )
-                )
-
-                totalCovered += covered
-                totalExecutable += executable
-            }
-        }
-
-        guard !fileCoverages.isEmpty else {
-            return nil
-        }
-
-        let overallCoverage = totalExecutable > 0 ? (Double(totalCovered) / Double(totalExecutable)) * 100.0 : 0.0
-
-        return CodeCoverage(lineCoverage: overallCoverage, files: fileCoverages)
-    }
-
-    private static func parseSPMFormat(json: [String: Any]) -> CodeCoverage? {
-        guard let dataArray = json["data"] as? [[String: Any]],
-            let firstData = dataArray.first,
-            let filesArray = firstData["files"] as? [[String: Any]]
-        else {
-            return nil
-        }
-
-        var fileCoverages: [FileCoverage] = []
-        var totalCovered = 0
-        var totalExecutable = 0
-
-        for fileData in filesArray {
-            guard let filename = fileData["filename"] as? String,
-                let summary = fileData["summary"] as? [String: Any],
-                let lines = summary["lines"] as? [String: Any],
-                let covered = lines["covered"] as? Int,
-                let count = lines["count"] as? Int
-            else {
-                continue
-            }
-
-            let coverage = count > 0 ? (Double(covered) / Double(count)) * 100.0 : 0.0
-            let name = (filename as NSString).lastPathComponent
-
-            fileCoverages.append(
-                FileCoverage(
-                    path: filename,
-                    name: name,
-                    lineCoverage: coverage,
-                    coveredLines: covered,
-                    executableLines: count
-                )
-            )
-
-            totalCovered += covered
-            totalExecutable += count
-        }
-
-        guard !fileCoverages.isEmpty else {
-            return nil
-        }
-
-        let overallCoverage = totalExecutable > 0 ? (Double(totalCovered) / Double(totalExecutable)) * 100.0 : 0.0
-
-        return CodeCoverage(lineCoverage: overallCoverage, files: fileCoverages)
     }
 }
